% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Create a function to coerce and combine two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Binding Two factors
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
location_a <- factor(c("Pilani", "Hyderabad"))
location_b <-  c("Goa", "Dubai")
location <- fbind(location_a, location_b)
}
